Move AbstractWorker to its own module
