+++ bot/multithreading/worker/queue.py
@@ -1,7 +1,7 @@
 import queue
 
 from bot.multithreading.work import Work
-from bot.multithreading.worker import AbstractWorker
+from bot.multithreading.worker.abstract import AbstractWorker
 
 
 class QueueWorker(AbstractWorker):
